+++ toppra/algorithm/__init__.py
@@ -1,3 +1,5 @@
 """TOPP-RA implements multiple path-parametrization algorithms.
 """
 from .reachabilitybased import TOPPRA, TOPPRAsd
+
+__all__ = ["TOPPRA", "TOPPRAsd"]