+++ HISTORY.md
@@ -12,6 +12,7 @@ Major release! Implement TOPPRA in C++ and several improvements to Python codeba
 
 ### Changed
 
+- [python] Improve documentation for `toppra.constraint`
 - [python] #98: Eliminate use of deprecated method.
 - [cpp] Bug fixes in solver wrapper.
 - [python] Simplify TOPPRA class interface.